// Biblioteca/src/model/CategoriaLivro.ts
enum CategoriaLivro {
    FICCAO_CIENTIFICA = 'FICCAO_CIENTIFICA',
    ROMANCE = 'ROMANCE',
    FANTASIA = 'FANTASIA',
    BIOGRAFIA = 'BIOGRAFIA',
    HISTORIA = 'HISTORIA',
    TECNOLOGIA = 'TECNOLOGIA',
    CIENCIAS = 'CIENCIAS',
    LITERATURA = 'LITERATURA',
    AUTOAJUDA = 'AUTOAJUDA',
    EDUCACIONAL = 'EDUCACIONAL',
    INFANTIL = 'INFANTIL',
    POESIA = 'POESIA',
    MANGA = 'MANGA',
    QUADRINHOS = 'QUADRINHOS'
}

export default CategoriaLivro; 